requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
openai>=0.28.0
python-dotenv>=1.0.0